+++ estilos/style.css
@@ -66,24 +66,36 @@
     font-size: 50px;
     text-align: center;
     padding: 20px;
+    overflow: auto;
+   
+    
 }
 
 .containerTarea{
     border: 2px solid #0005;
-    padding: 10px;
+    display: flex;
+    margin: 20px;
+
 }
 
-
-
 .checkbox {
-    margin-right: 10px; 
+    margin-right: 3px; 
+    height: 50px;
+    width: 50px;
+    border-radius: 50px;
+    margin: auto;
+    margin-left: 12px;
 }
 
 .taskTexto {
+    margin-right: 500px;
     flex-grow: 1;
     text-align: center; 
+    font-size: 40px;
 }
 
 .icono {
     margin-left: 10px; 
+    height: 100px;
+    width: 100px;
 }
